feat: Migrated Search Module to KMP
