+++ feature/search/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/feature/search/SearchResultUiState.kt
