+++ feature/search/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/feature/search/SearchUiStatePreviewParameterProvider.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:Suppress("ktlint:standard:max-line-length")
+
+package com.google.samples.apps.nowinandroid.feature.search
+
+import com.google.samples.apps.nowinandroid.core.model.data.FollowableTopic
+import com.google.samples.apps.nowinandroid.core.ui.PreviewParameterData.newsResources
+import com.google.samples.apps.nowinandroid.core.ui.PreviewParameterData.topics
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameterProvider
+
+/**
+ * This [PreviewParameterProvider](https://developer.android.com/reference/kotlin/androidx/compose/ui/tooling/preview/PreviewParameterProvider)
+ * provides list of [SearchResultUiState] for Composable previews.
+ */
+class SearchUiStatePreviewParameterProvider : PreviewParameterProvider<SearchResultUiState> {
+    override val values: Sequence<SearchResultUiState> = sequenceOf(
+        SearchResultUiState.Success(
+            topics = topics.mapIndexed { i, topic ->
+                FollowableTopic(topic = topic, isFollowed = i % 2 == 0)
+            },
+            newsResources = newsResources,
+        ),
+    )
+}
